Fixed data size for open cell
Set to 16bit unsigned to match other cell voltages (8 bit is way too small)

+++ shepherd_bms/include/datastructs.h
@@ -20,7 +20,7 @@ struct ChipData_t
     //These are calculated during the analysis of data
     uint8_t cellTemp[NUM_CELLS_PER_CHIP];
     float cellResistance[NUM_CELLS_PER_CHIP];
-    uint8_t openCellVoltage[NUM_CELLS_PER_CHIP];
+    uint16_t openCellVoltage[NUM_CELLS_PER_CHIP];
 };
 
 /**
